Added calculator functionality

+++ app/src/main/java/com/example/tbc_midterm_project/presentation/event/calculator/CalculatorEvents.kt
@@ -1,6 +1,6 @@
 package com.example.tbc_midterm_project.presentation.event.calculator
 
 sealed class CalculatorEvents {
-    data class CalculatePressed(val sex: String, val age: Int, val height: Int, val weight: Int) :
+    data class CalculatePressed(val sex: String, val age: Int, val height: Double, val weight: Double) :
         CalculatorEvents()
 }
